backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
firebase-admin==6.7.0
PyJWT==2.10.1
